tweaked the fullSetEditor View
Working on this view has highlighted a major flaw in the RichTextField infastructure. I store the viewController that is presenting its contents in the RichTextFieldViewModels, but I now see that that will lead to a lot of unexpected behavior ( initially, before the RichTextFieldViewModel even existed this was to solve a stale reference problem that was happening in the UI ). I now see that ViewControllers should be intialized with text to display whenever, and should not be pulled from a model.